Solve day 8

+++ template/src/part1.rs
@@ -1,4 +1,4 @@
 #[allow(unused_variables)]
-pub fn solve(input: &str) -> i32 {
+pub fn solve(input: &str) -> i64 {
     0
 }
